linux/generic/3.7: add IPv6 NAT support and fixes.
This evil of a patch adds kernel support for IPv6 NAT under the 3.7
series kernel. If you are anyone who ships devices commercially, you are
kindly requested NOT to include this code in your image; the sole
purpose of this code should be to provide a means of getting around
geolocation restrictions.

I hope nobody sane wants to see all the stupid and lazy ISPs out there
dishing out single on-link IPv6 addresses, forcing customers to use NAT.
Just Say No.

In order to actually make use of what is provided in this patch, it is
required to also have iptables 1.4.17 or higher.

Signed-off-by: Oliver Smith <olipro@8.c.9.b.0.7.4.0.1.0.0.2.ip6.arpa>